// React/app/GerAdv_TS/Fornecedores/Interfaces/interface.Fornecedores.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface IFornecedores {
// 202501251
    id: number;
 
	cidade: number,
	grupo: number,
	nome: string,
	subgrupo: number,
	tipo: boolean,
	sexo: boolean,
	cnpj: string,
	inscest: string,
	cpf: string,
	rg: string,
	endereco: string,
	bairro: string,
	cep: string,
	fone: string,
	fax: string,
	email: string,
	site: string,
	obs: string,
	produtos: string,
	contatos: string,
}

export interface IFornecedoresFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IFornecedoresIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}